Remove backend files
I initially thought that hosting backend and frontend code
together would be more convinient, for the sake maintenance.
However, since the backend and frontend will be hosted on two separate
service, with different lifecycles. It makes sense to keep
the codebases independent. This does mean an additional step
in terms of coding. But it does frontend the advantage of developing
for the live version of the api. Backend guys, I hope you're ready
for the challenge. (It's not that big a deal, you just develop for
frontend/master)e. (It's not that big a deal, you just develop for